ovn: support acl logging for INVALID packets
INVALID packets doesn't map to any ACL, but sometimes we need to
know when these packets are detected and dropped. This patch supports
logging for such packets and provides a global configuration in NB
to enable/disable this behavior. Default behavior is not logging,
so that it is backward compatible.